updated topo2interaction to better handle weights
Previously, the code was written with the intention of weights always being between -1 and 1. But it was imperfect, so modified it a little. After the modification, the standard topo files with two type value either 1 or 2 are modified to treat 2 as -1. More than 2 type values are always treated as float point weights. The case where there are only two type values which are not -1,1 or 1,2 might end up giving weird results, but nothing can be done there.